Use a more appropriate type for tail_start.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@55434 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/algo/blast/api/seedtop.hpp
@@ -54,7 +54,7 @@ struct SPatternUnit {
     size_t at_most;
     bool is_x;
     SPatternUnit(const string unit) {
-        unsigned int tail_start = 0;
+        size_t tail_start = 0;
         bool parse_failed = false;
         is_x = false;
         switch(unit[0]) {
